move maven models to shared and deserialization in backend
