+++ rest-lib-utils/src/main/java/com/sap/psr/vulas/cia/util/ArtifactDownloader.java
@@ -44,9 +44,9 @@ import org.springframework.web.client.RequestCallback;
 import org.springframework.web.client.ResponseExtractor;
 import org.springframework.web.client.RestTemplate;
 
-import com.sap.psr.vulas.cia.model.mavenCentral.MavenVersionsSearch;
-import com.sap.psr.vulas.cia.model.mavenCentral.ResponseDoc;
 import com.sap.psr.vulas.shared.json.model.Artifact;
+import com.sap.psr.vulas.shared.json.model.mavenCentral.MavenVersionsSearch;
+import com.sap.psr.vulas.shared.json.model.mavenCentral.ResponseDoc;
 import com.sap.psr.vulas.shared.util.VulasConfiguration;
 
 /**
